Rename & refactor helpers.
